+++ ChangeLog.md
@@ -14,6 +14,8 @@ BLOB
 * Added support for large block blob. 
 * Added `publicAccessLevel` to `ContainerResult` for the APIs `listContainersSegmented` and `listContainersSegmentedWithPrefix`.
 * When specifiying access condition `If-None-Match: *` for reading, it will always fail.
+* Returned content MD5 for range gets Blobs.
+* Fixed the issue that `useTransactionalMD5` didn't take effect for downloading a big blob.
 
 QUEUE
 * `createMessage` callback has been changed from `errorOrResponse` to `errorOrResult<QueueMessageResult>` which contains `messageId`, `popReceipt`, `timeNextVisible`, `insertionTime` and `expirationTime`. It can be passed to updateMessage and deleteMessage APIs.